+++ src/main/java/com/org/bgv/common/EmployerPackageDocumentResponse.java
@@ -1,5 +1,7 @@
 package com.org.bgv.common;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -14,7 +16,7 @@ public class EmployerPackageDocumentResponse {
     private Long id;
     private CategoryInfo checkCategory;
     private DocumentTypeInfo documentType;
-    private Double addonPrice;
+    private BigDecimal addonPrice;
     private Boolean includedInBase;
     private String selectionType;
 }
