+++ src/main/java/com/org/bgv/common/VerificationCaseResponse.java
@@ -1,5 +1,6 @@
 package com.org.bgv.common;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -18,9 +19,9 @@ public class VerificationCaseResponse {
     private Long candidateId;
     private Long companyId;
     private EmployerPackageInfo employerPackage;
-    private Double basePrice;
-    private Double addonPrice;
-    private Double totalPrice;
+    private BigDecimal basePrice;
+    private BigDecimal addonPrice;
+    private BigDecimal totalPrice;
     private String status;
     private LocalDateTime createdAt;
     private List<VerificationCaseDocumentResponse> documents;
